moved some more stuff out of index.js and into helper files in the js directory.
